feat: wordList 주어진 단어 검증 로직 작성
