+++ src/test/java/wordle/domain/Word.java
@@ -1,0 +1,24 @@
+package wordle.domain;
+
+import java.util.Objects;
+
+public class Word {
+    private String word;
+
+    public Word(String word) {
+        this.word = word;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Word word1 = (Word) o;
+        return Objects.equals(word, word1.word);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(word);
+    }
+}
