+++ src/test/java/wordle/domain/WordList.java
@@ -1,35 +1,15 @@
 package wordle.domain;
 
-import org.junit.jupiter.api.Test;
+import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+public class WordList {
+    private List<Word> wordList;
 
-public class WordListTest {
-
-    @Test
-    public void 주어진_단어가_wordList_안에_있으면_true를_반환한다() {
-        // given
-        Word word = new Word("cigar");
-        WordList wordList = new WordList();
-
-        // when
-        boolean actual = wordList.contains(word);
-
-        // then
-        assertTrue(actual);
+    public WordList(List<Word> wordList) {
+        this.wordList = wordList;
     }
 
-    @Test
-    public void 주어진_단어가_wordList_안에_없으면_false를_반환한다() {
-        // given
-        Word word = new Word("ooooo");
-        WordList wordList = new WordList();
-
-        // when
-        boolean actual = wordList.contains(word);
-
-        // then
-        assertFalse(actual);
+    public boolean contains(Word word) {
+        return wordList.contains(word);
     }
 }
